+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/Testnet.hs
@@ -80,6 +80,7 @@ import Ouroboros.Network.ConnectionManager.Test.Utils
            (abstractStateIsFinalTransition, connectionManagerTraceMap,
            validTransitionMap, verifyAbstractTransition,
            verifyAbstractTransitionOrder)
+import Ouroboros.Network.ConsensusMode
 import Ouroboros.Network.InboundGovernor.Test.Utils (inboundGovernorTraceMap,
            remoteStrIsFinalTransition, serverTraceMap, validRemoteTransitionMap,
            verifyRemoteTransition, verifyRemoteTransitionOrder)
@@ -629,22 +630,26 @@ prop_no_non_trustable_peers_before_caught_up_state ioSimTrace traceNumber =
     verify_no_non_trustable_peers_before_caught_up_state events =
       let govUseBootstrapPeers :: Signal UseBootstrapPeers
           govUseBootstrapPeers =
-            selectDiffusionPeerSelectionState (Governor.bootstrapPeersFlag) events
+            selectDiffusionPeerSelectionState (Governor.bootstrapPeersFlag)
+                                              events
 
           govLedgerStateJudgement :: Signal LedgerStateJudgement
           govLedgerStateJudgement =
-            selectDiffusionPeerSelectionState (Governor.ledgerStateJudgement) events
+            selectDiffusionPeerSelectionState (Governor.ledgerStateJudgement)
+                                              events
 
           govKnownPeers :: Signal (Set NtNAddr)
           govKnownPeers =
-            selectDiffusionPeerSelectionState (KnownPeers.toSet . Governor.knownPeers) events
+            selectDiffusionPeerSelectionState (KnownPeers.toSet . Governor.knownPeers)
+                                              events
 
           govTrustedPeers :: Signal (Set NtNAddr)
           govTrustedPeers =
             selectDiffusionPeerSelectionState
               (\st -> LocalRootPeers.keysSet (LocalRootPeers.clampToTrustable (Governor.localRootPeers st))
                    <> PublicRootPeers.getBootstrapPeers (Governor.publicRootPeers st)
-              ) events
+              )
+              events
 
           govHasOnlyBootstrapPeers :: Signal Bool
           govHasOnlyBootstrapPeers =
@@ -713,6 +718,7 @@ unit_4177 = prop_inbound_governor_transitions_coverage absNoAttenuation script
         (singletonTimedScript Map.empty)
         [ ( NodeArgs (-6) InitiatorAndResponderDiffusionMode (Just 180)
               (Map.fromList [(RelayAccessDomain "test2" 65535, DoAdvertisePeer)])
+              PraosMode
               (Script ((UseBootstrapPeers [RelayAccessDomain "bootstrap" 00000]) :| []))
               (TestAddress (IPAddr (read "0:7:0:7::") 65533))
               PeerSharingDisabled
@@ -720,15 +726,16 @@ unit_4177 = prop_inbound_governor_transitions_coverage absNoAttenuation script
               , (RelayAccessAddress "0:6:0:3:0:6:0:5" 65530,(DoNotAdvertisePeer, IsNotTrustable))])
               ]
               (Script (LedgerPools [] :| []))
-              nullPeerSelectionTargets {
-                targetNumberOfKnownPeers = 2,
-                targetNumberOfEstablishedPeers = 2,
-                targetNumberOfActivePeers = 1,
+              ConsensusModePeerTargets {
+                praosTargets = nullPeerSelectionTargets {
+                    targetNumberOfKnownPeers = 2,
+                    targetNumberOfEstablishedPeers = 2,
+                    targetNumberOfActivePeers = 1,
 
-                targetNumberOfKnownBigLedgerPeers = 0,
-                targetNumberOfEstablishedBigLedgerPeers = 0,
-                targetNumberOfActiveBigLedgerPeers = 0
-              }
+                    targetNumberOfKnownBigLedgerPeers = 0,
+                    targetNumberOfEstablishedBigLedgerPeers = 0,
+                    targetNumberOfActiveBigLedgerPeers = 0 },
+                genesisSyncTargets = nullPeerSelectionTargets }
               (Script (DNSTimeout {getDNSTimeout = 0.239} :| [DNSTimeout {getDNSTimeout = 0.181},DNSTimeout {getDNSTimeout = 0.185},DNSTimeout {getDNSTimeout = 0.14},DNSTimeout {getDNSTimeout = 0.221}]))
               (Script (DNSLookupDelay {getDNSLookupDelay = 0.067} :| [DNSLookupDelay {getDNSLookupDelay = 0.097},DNSLookupDelay {getDNSLookupDelay = 0.101},DNSLookupDelay {getDNSLookupDelay = 0.096},DNSLookupDelay {getDNSLookupDelay = 0.051}]))
               Nothing
@@ -744,17 +751,19 @@ unit_4177 = prop_inbound_governor_transitions_coverage absNoAttenuation script
           )
         , ( NodeArgs (1) InitiatorAndResponderDiffusionMode (Just 135)
              (Map.fromList [(RelayAccessAddress "0:7:0:7::" 65533, DoAdvertisePeer)])
+             PraosMode
               (Script ((UseBootstrapPeers [RelayAccessDomain "bootstrap" 00000]) :| []))
              (TestAddress (IPAddr (read "0:6:0:3:0:6:0:5") 65530))
              PeerSharingDisabled
              []
              (Script (LedgerPools [] :| []))
-             nullPeerSelectionTargets {
-               targetNumberOfRootPeers = 2,
-               targetNumberOfKnownPeers = 5,
-               targetNumberOfEstablishedPeers = 1,
-               targetNumberOfActivePeers = 1
-             }
+             ConsensusModePeerTargets {
+               praosTargets = nullPeerSelectionTargets {
+                   targetNumberOfRootPeers = 2,
+                   targetNumberOfKnownPeers = 5,
+                   targetNumberOfEstablishedPeers = 1,
+                   targetNumberOfActivePeers = 1 },
+               genesisSyncTargets = nullPeerSelectionTargets }
              (Script (DNSTimeout {getDNSTimeout = 0.28}
                   :| [DNSTimeout {getDNSTimeout = 0.204},
                       DNSTimeout {getDNSTimeout = 0.213}
@@ -811,7 +820,8 @@ prop_track_coolingToCold_demotions ioSimTracer traceNumber =
       <$> events
 
   where
-    verify_coolingToColdDemotions :: Events DiffusionTestTrace -> Property
+    verify_coolingToColdDemotions :: Events DiffusionTestTrace
+                                  -> Property
     verify_coolingToColdDemotions events =
       let govInProgressDemoteToCold :: Signal (Set NtNAddr)
           govInProgressDemoteToCold =
@@ -1301,6 +1311,7 @@ unit_4191 = testWithIOSim prop_diffusion_dns_can_recover 125000 absInfo script
             InitiatorAndResponderDiffusionMode
             (Just 224)
             Map.empty
+            PraosMode
             (Script ((UseBootstrapPeers [RelayAccessDomain "bootstrap" 00000]) :| []))
             (TestAddress (IPAddr (read "0.0.1.236") 65527))
             PeerSharingDisabled
@@ -1308,16 +1319,18 @@ unit_4191 = testWithIOSim prop_diffusion_dns_can_recover 125000 absInfo script
                                 , (RelayAccessDomain "test3" 4,(DoAdvertisePeer, IsNotTrustable))])
             ]
             (Script (LedgerPools [] :| []))
-            PeerSelectionTargets
-              { targetNumberOfRootPeers = 6,
-                targetNumberOfKnownPeers = 7,
-                targetNumberOfEstablishedPeers = 7,
-                targetNumberOfActivePeers = 6,
+            ConsensusModePeerTargets {
+              praosTargets = PeerSelectionTargets
+                { targetNumberOfRootPeers = 6,
+                  targetNumberOfKnownPeers = 7,
+                  targetNumberOfEstablishedPeers = 7,
+                  targetNumberOfActivePeers = 6,
 
-                targetNumberOfKnownBigLedgerPeers = 0,
-                targetNumberOfEstablishedBigLedgerPeers = 0,
-                targetNumberOfActiveBigLedgerPeers = 0
-              }
+                  targetNumberOfKnownBigLedgerPeers = 0,
+                  targetNumberOfEstablishedBigLedgerPeers = 0,
+                  targetNumberOfActiveBigLedgerPeers = 0
+                },
+              genesisSyncTargets = nullPeerSelectionTargets }
             (Script (DNSTimeout {getDNSTimeout = 0.406} :| [ DNSTimeout {getDNSTimeout = 0.11}
                                                            , DNSTimeout {getDNSTimeout = 0.333}
                                                            , DNSTimeout {getDNSTimeout = 0.352}
@@ -1396,7 +1409,8 @@ prop_diffusion_target_established_public ioSimTrace traceNumber =
         )
       <$> events
   where
-    verify_target_established_public :: Events DiffusionTestTrace -> Property
+    verify_target_established_public :: Events DiffusionTestTrace
+                                     -> Property
     verify_target_established_public events =
       let govPublicRootPeersSig :: Signal (Set NtNAddr)
           govPublicRootPeersSig =
@@ -1478,7 +1492,8 @@ prop_diffusion_target_active_public ioSimTrace traceNumber =
         )
       <$> events
   where
-    verify_target_active_public :: Events DiffusionTestTrace -> Property
+    verify_target_active_public :: Events DiffusionTestTrace
+                                -> Property
     verify_target_active_public events =
         let govPublicRootPeersSig :: Signal (Set NtNAddr)
             govPublicRootPeersSig =
@@ -1488,7 +1503,8 @@ prop_diffusion_target_active_public ioSimTrace traceNumber =
 
             govActivePeersSig :: Signal (Set NtNAddr)
             govActivePeersSig =
-              selectDiffusionPeerSelectionState Governor.activePeers events
+              selectDiffusionPeerSelectionState Governor.activePeers
+                                                events
 
             publicInActive :: Signal Bool
             publicInActive =
@@ -1548,16 +1564,19 @@ prop_diffusion_target_active_local ioSimTrace traceNumber =
         )
       <$> events
   where
-    verify_target_active_local :: Events DiffusionTestTrace -> Property
+    verify_target_active_local :: Events DiffusionTestTrace
+                               -> Property
     verify_target_active_local events =
         let govLocalRootPeersSig :: Signal (Set NtNAddr)
             govLocalRootPeersSig =
               selectDiffusionPeerSelectionState
-                (LocalRootPeers.keysSet . Governor.localRootPeers) events
+                (LocalRootPeers.keysSet . Governor.localRootPeers)
+                events
 
             govActivePeersSig :: Signal (Set NtNAddr)
             govActivePeersSig =
-              selectDiffusionPeerSelectionState Governor.activePeers events
+              selectDiffusionPeerSelectionState Governor.activePeers
+              events
 
             localInActive :: Signal Bool
             localInActive =
@@ -1619,7 +1638,8 @@ prop_diffusion_target_active_root ioSimTrace traceNumber =
         )
       <$> events
   where
-    verify_target_active_root :: Events DiffusionTestTrace -> Property
+    verify_target_active_root :: Events DiffusionTestTrace
+                              -> Property
     verify_target_active_root events =
         let govLocalRootPeersSig :: Signal (Set NtNAddr)
             govLocalRootPeersSig =
@@ -1718,7 +1738,7 @@ prop_diffusion_target_established_local ioSimTrace traceNumber =
                . Trace.take traceNumber
                $ ioSimTrace
 
-     in conjoin
+    in conjoin
       $ (\ev ->
         let evsList = eventsToList ev
             lastTime = fst
@@ -1731,8 +1751,9 @@ prop_diffusion_target_established_local ioSimTrace traceNumber =
       <$> events
 
   where
-    verify_target_established_local :: Events DiffusionTestTrace -> Property
-    verify_target_established_local events =
+    verify_target_established_local :: Events DiffusionTestTrace
+                                    -> Property
+    verify_target_established_local events  =
       let govLocalRootPeersSig :: Signal (LocalRootPeers.LocalRootPeers NtNAddr)
           govLocalRootPeersSig =
             selectDiffusionPeerSelectionState Governor.localRootPeers events
@@ -1896,7 +1917,8 @@ prop_diffusion_target_active_below ioSimTrace traceNumber =
       <$> events
 
   where
-    verify_target_active_below :: Events DiffusionTestTrace -> Property
+    verify_target_active_below :: Events DiffusionTestTrace
+                               -> Property
     verify_target_active_below events =
       let govLocalRootPeersSig :: Signal (LocalRootPeers.LocalRootPeers NtNAddr)
           govLocalRootPeersSig =
@@ -1964,7 +1986,7 @@ prop_diffusion_target_active_below ioSimTrace traceNumber =
 
           govInProgressPromoteWarmSig :: Signal (Set NtNAddr)
           govInProgressPromoteWarmSig =
-            selectDiffusionPeerSelectionState Governor.inProgressPromoteWarm events
+            selectDiffusionPeerSelectionState Governor.inProgressPromoteWarm  events
 
           trJoinKillSig :: Signal JoinedOrKilled
           trJoinKillSig =
@@ -2083,7 +2105,8 @@ prop_diffusion_target_active_local_below ioSimTrace traceNumber =
       <$> events
 
   where
-    verify_target_active_below :: Events DiffusionTestTrace -> Property
+    verify_target_active_below :: Events DiffusionTestTrace
+                               -> Property
     verify_target_active_below events =
       let govLocalRootPeersSig :: Signal (LocalRootPeers.LocalRootPeers NtNAddr)
           govLocalRootPeersSig =
@@ -2223,11 +2246,13 @@ async_demotion_network_script =
       (singletonTimedScript Map.empty)
       [ ( common { naAddr                  = addr1,
                    naLocalRootPeers        = localRoots1,
-                   naLocalSelectionTargets = Governor.nullPeerSelectionTargets {
-                                               targetNumberOfKnownPeers = 2,
-                                               targetNumberOfEstablishedPeers = 2,
-                                               targetNumberOfActivePeers = 2
-                                             }
+                   naPeerTargets = ConsensusModePeerTargets {
+                     praosTargets = Governor.nullPeerSelectionTargets {
+                         targetNumberOfKnownPeers = 2,
+                           targetNumberOfEstablishedPeers = 2,
+                           targetNumberOfActivePeers = 2
+                         },
+                     genesisSyncTargets = peerTargets }
                  }
         , [ JoinNetwork 0
             -- reconfigure the peer to trigger the outbound governor log
@@ -2261,21 +2286,24 @@ async_demotion_network_script =
         saSlot             = secondsToDiffTime 1,
         saQuota            = 5  -- 5% chance of producing a block
       }
+    peerTargets = Governor.nullPeerSelectionTargets {
+      targetNumberOfKnownPeers = 1,
+      targetNumberOfEstablishedPeers = 1,
+      targetNumberOfActivePeers =1 }
+
     common = NodeArgs {
         naSeed             = 10,
         naDiffusionMode    = InitiatorAndResponderDiffusionMode,
         naMbTime           = Just 1,
         naPublicRoots      = Map.empty,
+        naConsensusMode    = PraosMode,
         naBootstrapPeers   = (Script ((UseBootstrapPeers [RelayAccessDomain "bootstrap" 00000]) :| [])),
         naAddr             = undefined,
         naLocalRootPeers   = undefined,
         naLedgerPeers      = Script (LedgerPools [] :| []),
-        naLocalSelectionTargets
-                           = Governor.nullPeerSelectionTargets {
-                               targetNumberOfKnownPeers = 1,
-                               targetNumberOfEstablishedPeers = 1,
-                               targetNumberOfActivePeers = 1
-                             },
+        naPeerTargets      = ConsensusModePeerTargets {
+            praosTargets = peerTargets,
+            genesisSyncTargets = peerTargets },
         naDNSTimeoutScript = singletonScript (DNSTimeout 3),
         naDNSLookupDelayScript
                            = singletonScript (DNSLookupDelay 0.2),
@@ -2459,12 +2487,13 @@ prop_diffusion_target_active_local_above ioSimTrace traceNumber =
                      $ evsList
          in classifySimulatedTime lastTime
           $ classifyNumberOfEvents (length evsList)
-          $ verify_target_active_above ev
+          $ verify_target_active_above ev 
         )
       <$> events
 
   where
-    verify_target_active_above :: Events DiffusionTestTrace -> Property
+    verify_target_active_above :: Events DiffusionTestTrace
+                               -> Property
     verify_target_active_above events =
       let govLocalRootPeersSig :: Signal (LocalRootPeers.LocalRootPeers NtNAddr)
           govLocalRootPeersSig =
@@ -2704,17 +2733,19 @@ prop_unit_4258 =
         (singletonTimedScript Map.empty)
         [( NodeArgs (-3) InitiatorAndResponderDiffusionMode (Just 224)
              (Map.fromList [])
+             PraosMode
              (Script ((UseBootstrapPeers [RelayAccessDomain "bootstrap" 00000]) :| []))
              (TestAddress (IPAddr (read "0.0.0.4") 9))
              PeerSharingDisabled
              [(1,1,Map.fromList [(RelayAccessAddress "0.0.0.8" 65531,(DoNotAdvertisePeer, IsNotTrustable))])]
              (Script (LedgerPools [] :| []))
-             nullPeerSelectionTargets {
+             ConsensusModePeerTargets {
+               praosTargets = nullPeerSelectionTargets {
                  targetNumberOfRootPeers = 2,
                  targetNumberOfKnownPeers = 5,
                  targetNumberOfEstablishedPeers = 4,
-                 targetNumberOfActivePeers = 1
-               }
+                 targetNumberOfActivePeers = 1 },
+               genesisSyncTargets = nullPeerSelectionTargets }
              (Script (DNSTimeout {getDNSTimeout = 0.397}
                  :| [ DNSTimeout {getDNSTimeout = 0.382},
                       DNSTimeout {getDNSTimeout = 0.321},
@@ -2737,12 +2768,14 @@ prop_unit_4258 =
          ),
          ( NodeArgs (-5) InitiatorAndResponderDiffusionMode (Just 269)
              (Map.fromList [(RelayAccessAddress "0.0.0.4" 9, DoAdvertisePeer)])
+             PraosMode
              (Script ((UseBootstrapPeers [RelayAccessDomain "bootstrap" 00000]) :| []))
              (TestAddress (IPAddr (read "0.0.0.8") 65531))
              PeerSharingDisabled
              [(1,1,Map.fromList [(RelayAccessAddress "0.0.0.4" 9,(DoNotAdvertisePeer, IsNotTrustable))])]
              (Script (LedgerPools [] :| []))
-             nullPeerSelectionTargets {
+             ConsensusModePeerTargets {
+               praosTargets = nullPeerSelectionTargets {
                  targetNumberOfRootPeers = 4,
                  targetNumberOfKnownPeers = 5,
                  targetNumberOfEstablishedPeers = 3,
@@ -2751,7 +2784,8 @@ prop_unit_4258 =
                  targetNumberOfKnownBigLedgerPeers = 0,
                  targetNumberOfEstablishedBigLedgerPeers = 0,
                  targetNumberOfActiveBigLedgerPeers = 0
-               }
+               },
+               genesisSyncTargets = nullPeerSelectionTargets }
              (Script (DNSTimeout {getDNSTimeout = 0.281}
                  :| [ DNSTimeout {getDNSTimeout = 0.177},
                       DNSTimeout {getDNSTimeout = 0.164},
@@ -2807,6 +2841,7 @@ prop_unit_reconnect =
               InitiatorAndResponderDiffusionMode
               (Just 224)
               Map.empty
+              PraosMode
               (Script (DontUseBootstrapPeers :| []))
               (TestAddress (IPAddr (read "0.0.0.0") 0))
               PeerSharingDisabled
@@ -2815,16 +2850,17 @@ prop_unit_reconnect =
                                   ])
               ]
               (Script (LedgerPools [] :| []))
-              PeerSelectionTargets {
-                targetNumberOfRootPeers = 1,
-                targetNumberOfKnownPeers = 1,
-                targetNumberOfEstablishedPeers = 1,
-                targetNumberOfActivePeers = 1,
+              ConsensusModePeerTargets {
+                praosTargets = PeerSelectionTargets {
+                    targetNumberOfRootPeers = 1,
+                    targetNumberOfKnownPeers = 1,
+                    targetNumberOfEstablishedPeers = 1,
+                    targetNumberOfActivePeers = 1,
 
-                targetNumberOfKnownBigLedgerPeers = 0,
-                targetNumberOfEstablishedBigLedgerPeers = 0,
-                targetNumberOfActiveBigLedgerPeers = 0
-              }
+                    targetNumberOfKnownBigLedgerPeers = 0,
+                    targetNumberOfEstablishedBigLedgerPeers = 0,
+                    targetNumberOfActiveBigLedgerPeers = 0 },
+                genesisSyncTargets = nullPeerSelectionTargets }
               (Script (DNSTimeout {getDNSTimeout = 10} :| []))
               (Script (DNSLookupDelay {getDNSLookupDelay = 0} :| []))
               Nothing
@@ -2837,21 +2873,23 @@ prop_unit_reconnect =
                InitiatorAndResponderDiffusionMode
                (Just 2)
                Map.empty
+               PraosMode
                (Script (DontUseBootstrapPeers :| []))
                (TestAddress (IPAddr (read "0.0.0.1") 0))
                PeerSharingDisabled
                [(1,1,Map.fromList [(RelayAccessAddress "0.0.0.0" 0,(DoNotAdvertisePeer, IsNotTrustable))])]
                (Script (LedgerPools [] :| []))
-               PeerSelectionTargets {
-                 targetNumberOfRootPeers = 1,
-                 targetNumberOfKnownPeers = 1,
-                 targetNumberOfEstablishedPeers = 1,
-                 targetNumberOfActivePeers = 1,
+               ConsensusModePeerTargets {
+                 praosTargets = PeerSelectionTargets {
+                     targetNumberOfRootPeers = 1,
+                     targetNumberOfKnownPeers = 1,
+                     targetNumberOfEstablishedPeers = 1,
+                     targetNumberOfActivePeers = 1,
 
-                 targetNumberOfKnownBigLedgerPeers = 0,
-                 targetNumberOfEstablishedBigLedgerPeers = 0,
-                 targetNumberOfActiveBigLedgerPeers = 0
-               }
+                     targetNumberOfKnownBigLedgerPeers = 0,
+                     targetNumberOfEstablishedBigLedgerPeers = 0,
+                     targetNumberOfActiveBigLedgerPeers = 0 },
+                 genesisSyncTargets = nullPeerSelectionTargets }
              (Script (DNSTimeout {getDNSTimeout = 10} :| [ ]))
              (Script (DNSLookupDelay {getDNSLookupDelay = 0} :| []))
              Nothing
@@ -3237,17 +3275,18 @@ unit_peer_sharing =
     ip_2 = TestAddress $ IPAddr (IP.IPv4 (IP.toIPv4 [0,0,0,0])) 3002
     -- ra_2 = RelayAccessAddress (IP.IPv4 (IP.toIPv4 [0,0,0,0])) 3002
 
-    targets x = PeerSelectionTargets {
-        targetNumberOfRootPeers = x,
-        targetNumberOfKnownPeers = x,
-        targetNumberOfEstablishedPeers = x,
-        targetNumberOfActivePeers = x,
-        targetNumberOfKnownBigLedgerPeers = 0,
-        targetNumberOfEstablishedBigLedgerPeers = 0,
-        targetNumberOfActiveBigLedgerPeers = 0
-      }
+    targets x = let t = PeerSelectionTargets {
+                          targetNumberOfRootPeers = x,
+                          targetNumberOfKnownPeers = x,
+                          targetNumberOfEstablishedPeers = x,
+                          targetNumberOfActivePeers = x,
+                          targetNumberOfKnownBigLedgerPeers = 0,
+                          targetNumberOfEstablishedBigLedgerPeers = 0,
+                          targetNumberOfActiveBigLedgerPeers = 0 }
+                in ConsensusModePeerTargets { praosTargets = t, genesisSyncTargets = t }
 
-    defaultNodeArgs = NodeArgs {
+
+    defaultNodeArgs naConsensusMode = NodeArgs {
         naSeed = 0,
         naDiffusionMode = InitiatorAndResponderDiffusionMode,
         naMbTime = Nothing,
@@ -3257,32 +3296,35 @@ unit_peer_sharing =
         naPeerSharing = PeerSharingEnabled,
         naLocalRootPeers = undefined,
         naLedgerPeers = singletonScript (LedgerPools []),
-        naLocalSelectionTargets = undefined,
+        naPeerTargets = ConsensusModePeerTargets {
+            praosTargets = nullPeerSelectionTargets,
+            genesisSyncTargets = nullPeerSelectionTargets },
         naDNSTimeoutScript = singletonScript (DNSTimeout 300),
         naDNSLookupDelayScript = singletonScript (DNSLookupDelay 0.01),
         naChainSyncEarlyExit = False,
         naChainSyncExitOnBlockNo = Nothing,
-        naFetchModeScript = singletonScript FetchModeDeadline
+        naFetchModeScript = singletonScript FetchModeDeadline,
+        naConsensusMode
       }
 
     script = DiffusionScript
                (mainnetSimArgs 3)
                (singletonScript (mempty, ShortDelay))
-               [ ( defaultNodeArgs { naAddr = ip_0,
+               [ ( (defaultNodeArgs GenesisMode) { naAddr = ip_0,
                                      naLocalRootPeers = [(1, 1, Map.fromList [(ra_1, (DoNotAdvertisePeer, IsNotTrustable))])],
-                                     naLocalSelectionTargets = targets 1
+                                     naPeerTargets = targets 1
                                    }
                  , [JoinNetwork 0]
                  )
-               , ( defaultNodeArgs { naAddr = ip_1,
+               , ( (defaultNodeArgs PraosMode) { naAddr = ip_1,
                                      naLocalRootPeers = [],
-                                     naLocalSelectionTargets = targets 2
+                                     naPeerTargets = targets 2
                                    }
                  , [JoinNetwork 0]
                  )
-               , ( defaultNodeArgs { naAddr = ip_2,
+               , ( (defaultNodeArgs GenesisMode) { naAddr = ip_2,
                                      naLocalRootPeers = [(1, 1, Map.fromList [(ra_1, (DoNotAdvertisePeer, IsNotTrustable))])],
-                                     naLocalSelectionTargets = targets 2
+                                     naPeerTargets = targets 2
                                    }
                  , [JoinNetwork 0]
                  )
@@ -3296,9 +3338,9 @@ prop_churn_notimeouts :: SimTrace Void
                       -> Int
                       -> Property
 prop_churn_notimeouts ioSimTrace traceNumber =
-    let events :: [Events DiffusionTestTrace]
-        events = Trace.toList
-               . fmap ( Signal.eventsFromList
+   let events :: [Events DiffusionTestTrace]
+       events = Trace.toList
+                . fmap ( Signal.eventsFromList
                       . fmap (\(WithName _ (WithTime t b)) -> (t, b))
                       )
                . splitWithNameTrace
@@ -3677,11 +3719,23 @@ selectDiffusionPeerSelectionState :: Eq a
 selectDiffusionPeerSelectionState f =
     Signal.nub
   -- TODO: #3182 Rng seed should come from quickcheck.
-  . Signal.fromChangeEvents (f $ Governor.emptyPeerSelectionState (mkStdGen 42))
+  . (\evs ->
+       let evsList = Signal.eventsToList evs
+       in
+         case evsList of
+           [] -> Signal.fromChangeEvents (initialState PraosMode) (snd <$> evs)
+           (_, (consensusMode, _)):_ ->
+             Signal.fromChangeEvents (initialState consensusMode) (snd <$> evs)
+    )
   . Signal.selectEvents
       (\case
-        DiffusionDebugPeerSelectionTrace (TraceGovernorState _ _ st) -> Just (f st)
+        DiffusionDebugPeerSelectionTrace (TraceGovernorState _ _ st) -> Just (Governor.consensusMode st, f st)
         _                                                            -> Nothing)
+  where
+    initialState consensusMode =
+      f $ Governor.emptyPeerSelectionState
+            (mkStdGen 42)
+            consensusMode
 
 selectDiffusionPeerSelectionState' :: Eq a
                                   => (forall peerconn. Governor.PeerSelectionState NtNAddr peerconn -> a)
@@ -3689,7 +3743,7 @@ selectDiffusionPeerSelectionState' :: Eq a
                                   -> Signal a
 selectDiffusionPeerSelectionState' f =
   -- TODO: #3182 Rng seed should come from quickcheck.
-    Signal.fromChangeEvents (f $ Governor.emptyPeerSelectionState (mkStdGen 42))
+    Signal.fromChangeEvents (f $ Governor.emptyPeerSelectionState (mkStdGen 42) PraosMode)
   . Signal.selectEvents
       (\case
         DiffusionDebugPeerSelectionTrace (TraceGovernorState _ _ st) -> Just (f st)
@@ -3788,9 +3842,9 @@ labelDiffusionScript (DiffusionScript args _ nodes) =
     . label ("Nº nodes in InitiatorOnlyDiffusionMode: "
               ++ show (length $ filter ((== InitiatorOnlyDiffusionMode) . naDiffusionMode . fst) $ nodes))
     . label ("Nº active peers: "
-              ++ show (sum . map (targetNumberOfActivePeers . naLocalSelectionTargets . fst) $ nodes))
+              ++ show (sum . map (targetNumberOfActivePeers . praosTargets . naPeerTargets . fst) $ nodes))
     . label ("Nº active big ledger peers: "
-              ++ show (sum . map (targetNumberOfActiveBigLedgerPeers . naLocalSelectionTargets . fst) $ nodes))
+              ++ show (sum . map (targetNumberOfActiveBigLedgerPeers . praosTargets . naPeerTargets . fst) $ nodes))
     . label ("average number of active local roots: "
               ++ show (average . map (sum . map (\(HotValency v,_,_) -> v) . naLocalRootPeers . fst) $ nodes))
   where